requirements.txt
pyTelegramBotAPI>=4.14.0
rapidfuzz>=3.5.0
openpyxl>=3.1.2
flask>=3.0.0
flask-socketio>=5.3.6
flask-login>=0.6.3
simple-websocket>=1.0.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
Werkzeug>=3.0.0